requirements.txt
tqdm
numpy
opencv-python
pillow
pandas
pyocr
